机器学习-图灵/Project - ClassifyLeaves/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.65.0
pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
kaggle>=1.5.0 
timm>=0.9.0